exception logic added and token refresh logic optimized

+++ src/controllers/UserController.js
@@ -1,42 +1,79 @@
 const User = require("../models/User");
+const { StatusCodes } = require("http-status-codes");
+const Exception = require("../exceptions/exceptions");
 
-const getUsers = async (req, res) => {
+const getUsers = async (req, res, next) => {
   try {
     const users = await User.find();
-    res.status(200).json({ users });
+    if (!users) {
+      throw new Exception({
+        code: StatusCodes.NOT_FOUND,
+        message: "No users found",
+      });
+    }
+    res.status(StatusCodes.OK).json({ success: true, users });
   } catch (error) {
-    res.status(400).send(error.message);
+    next(error);
   }
 };
 
-const getUser = async (req, res) => {
+const getUser = async (req, res, next) => {
   try {
     const { username } = req.params;
     const user = await User.findOne({ username: username });
-    console.log(user);
-    res.status(200).json({ status: "success", user });
+    if (!user) {
+      throw new Exception({
+        code: StatusCodes.NOT_FOUND,
+        message: "User not found",
+      });
+    }
+    res.status(200).json({ success: true, user: user.username });
   } catch (error) {
-    res.status(400).send(error.message);
+    next(error);
   }
 };
 
-const updateUser = async (req, res) => {
+const updateUser = async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      throw new Exception({
+        code: StatusCodes.BAD_REQUEST,
+        message: "User ID is required to update user",
+      });
+    }
     const user = await User.findByIdAndUpdate(id, req.body);
-    res.status(200).json({ user });
+    if (!user) {
+      throw new Exception({
+        code: StatusCodes.NOT_FOUND,
+        message: "User not found",
+      });
+    }
+    res.status(200).json({ status: true, user });
   } catch (error) {
-    res.status(400).send(error.message);
+    next(error);
   }
 };
 
 const deleteUser = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      throw new Exception({
+        code: StatusCodes.BAD_REQUEST,
+        message: "User ID is required to delete user",
+      });
+    }
     const user = await User.findByIdAndDelete(id);
-    res.status(200).json({ user });
+    if (!user) {
+      throw new Exception({
+        code: StatusCodes.NOT_FOUND,
+        message: "User not found",
+      });
+    }
+    res.status(StatusCodes.OK).json({ status: true, user });
   } catch (error) {
-    res.status(400).send(error.message);
+    next(error);
   }
 };
